+++ packages/cloud-cognitive/src/components/Datagrid/useOnRowClick.js
@@ -1,7 +1,7 @@
 /*
  * Licensed Materials - Property of IBM
  * 5724-Q36
- * (c) Copyright IBM Corp. 2020
+ * (c) Copyright IBM Corp. 2020, 2023
  * US Government Users Restricted Rights - Use, duplication or disclosure
  * restricted by GSA ADP Schedule Contract with IBM Corp.
  */
@@ -12,9 +12,9 @@ const useOnRowClick = (hooks) => {
     const getRowProps = (props, datagridState) => {
       const { isFetching, row, instance } = datagridState;
       const { id, toggleRowSelected } = row;
-      const onClick = () => {
+      const onClick = (event) => {
         if (!isFetching && onRowClick) {
-          onRowClick(row);
+          onRowClick(row, event);
           instance.selectedFlatRows &&
             instance.selectedFlatRows.map((toggleRow) =>
               toggleRow.toggleRowSelected(false)
@@ -23,10 +23,18 @@ const useOnRowClick = (hooks) => {
         }
       };
 
+      const onKeyDown = (event) => {
+        const { key } = event;
+        if (key === 'Enter') {
+          onClick();
+        }
+      };
+
       return [
         props,
-        { onClick },
+        { onClick, onKeyDown },
         {
+          tabIndex: 0,
           style: {
             cursor: 'pointer',
           },