+++ packages/cloud-cognitive/src/components/index.js
@@ -74,6 +74,7 @@ export {
   useColumnOrder,
   useInlineEdit,
   useFiltering,
+  getAutoSizedColumnWidth,
 } from './Datagrid';
 export { EditTearsheet } from './EditTearsheet';
 export { EditTearsheetNarrow } from './EditTearsheetNarrow';
@@ -81,3 +82,5 @@ export { EditFullPage } from './EditFullPage';
 export { EditUpdateCards } from './EditUpdateCards';
 
 export { InlineEdit } from './InlineEdit';
+export { EditInPlace } from './EditInPlace';
+export { NonLinearReading } from './NonLinearReading';